+++ src/App/GeoFeatureGroupExtension.cpp
@@ -30,7 +30,7 @@
 #include "GeoFeatureGroupExtension.h"
 #include "Link.h"
 #include "Origin.h"
-#include "OriginFeature.h"
+#include "Datums.h"
 #include "OriginGroupExtension.h"
