+++ src/App/Origin.h
@@ -25,8 +25,8 @@
 #ifndef APP_Origin_H
 #define APP_Origin_H
 
+#include "Datums.h"
 #include "GeoFeatureGroupExtension.h"
-#include "OriginFeature.h"
 
 #include "QCoreApplication"
