+++ src/Mod/Part/App/Attacher.cpp
@@ -61,7 +61,7 @@
 
 #include <App/Application.h>
 #include <App/Document.h>
-#include <App/OriginFeature.h>
+#include <App/Datums.h>
 #include <Base/Console.h>
 
 #include "Attacher.h"
