+++ src/Mod/Part/Gui/TaskAttacher.cpp
@@ -34,7 +34,7 @@
 #include <App/Application.h>
 #include <App/Document.h>
 #include <App/ObjectIdentifier.h>
-#include <App/OriginFeature.h>
+#include <App/Datums.h>
 #include <App/Part.h>
 #include <Gui/Application.h>
 #include <Gui/BitmapFactory.h>
