+++ src/Mod/PartDesign/App/FeatureLinearPattern.cpp
@@ -32,7 +32,7 @@
 # include <TopoDS_Face.hxx>
 #endif
 
-#include <App/OriginFeature.h>
+#include <App/Datums.h>
 #include <Base/Axis.h>
 #include <Mod/Part/App/TopoShape.h>
 #include <Mod/Part/App/Part2DObject.h>
