+++ src/Mod/PartDesign/Gui/SketchWorkflow.cpp
@@ -47,7 +47,7 @@
 #include <App/Document.h>
 #include <App/Link.h>
 #include <App/Origin.h>
-#include <App/OriginFeature.h>
+#include <App/Datums.h>
 #include <App/Part.h>
 #include <Gui/Application.h>
 #include <Gui/Command.h>
