+++ src/Mod/PartDesign/Gui/TaskFeaturePick.cpp
@@ -31,7 +31,7 @@
 
 #include <App/Document.h>
 #include <App/Origin.h>
-#include <App/OriginFeature.h>
+#include <App/Datums.h>
 #include <App/Part.h>
 #include <Base/Console.h>
 #include <Gui/Application.h>
