+++ src/Mod/PartDesign/Gui/Utils.cpp
@@ -29,7 +29,7 @@
 #endif
 
 #include <App/Origin.h>
-#include <App/OriginFeature.h>
+#include <App/Datums.h>
 #include <App/Part.h>
 #include <Gui/Application.h>
 #include <Gui/CommandT.h>
