+++ src/Mod/Sketcher/App/SketchObject.cpp
@@ -87,7 +87,7 @@
 #include <App/IndexedName.h>
 #include <App/MappedName.h>
 #include <App/ObjectIdentifier.h>
-#include <App/OriginFeature.h>
+#include <App/Datums.h>
 #include <App/Part.h>
 #include <Base/Console.h>
 #include <Base/Reader.h>
