+++ src/Mod/Sketcher/Gui/Command.cpp
@@ -34,7 +34,7 @@
 #endif
 
 #include <App/DocumentObjectGroup.h>
-#include <App/OriginFeature.h>
+#include <App/Datums.h>
 #include <Gui/Action.h>
 #include <Gui/Application.h>
 #include <Gui/BitmapFactory.h>
